+++ packages/tailor-ui/src/Radio/RadioContext.ts
@@ -3,8 +3,8 @@ import { createContext } from 'react';
 export type Direction = 'horizontal' | 'vertical';
 
 const RadioContext = createContext<{
-  _onChange?: (value: string) => void;
-  _isChecked?: (value: string) => boolean;
+  handleChange?: (value: string) => void;
+  isChecked?: (value: string) => boolean;
   direction: Direction;
 }>({
   direction: 'horizontal',
